taint/test/mochitest/test_xhr_response_into_sink.html

<!DOCTYPE HTML>
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <meta charset="utf-8">
    <title>Check XMLHttpRequest Respones Source</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script "text/javascript">

      function runTest(method, r, f) {
            let request = new XMLHttpRequest();
            request.onload = () => {
                let resp = r(request);
                SimpleTest.info(`done! Response: ${resp}`);
                check_tainted(resp);
                f(resp);

            };
            //request.open("GET", "http://mochi.test:8888/tests/SimpleTest/setup.js", true);
            request.open("GET", `http://mochi.test:8888/tests/taint/test/mochitest/payload.sjs?${method}`, true);
            //request.open("GET", "/tests/SimpleTest/test.css", true);
            request.send("");
        }

        function runTest2(method, r, f, responseType="text") {
            let request = new XMLHttpRequest();
            request.responseType = responseType;
            request.onload = () => {
                let resp = r(request);
                SimpleTest.info(`done! Response: ${resp}`);
                f(resp);

            };
            //request.open("GET", "http://mochi.test:8888/tests/SimpleTest/setup.js", true);
            request.open("GET", `http://mochi.test:8888/tests/taint/test/mochitest/xhr_server.sjs?${method}`, true);
            //request.open("GET", "/tests/SimpleTest/test.css", true);
            request.send("");
        }


      let i = 0;
              let sink_names = [
        "eval",
        "innerHTML",
        "document.writeln",
        "document.write",
                        "innerHTML",


                        "innerHTML",
        "innerHTML",

      ];
      let contents = [

        'console.log("remote loaded payload");'
        ,"hello!"
        ,"hello!"
        ,"hello!"
        ,"<b>hi</b>"
        ,"hello"
        ,"hello"
      ];

      SimpleTest.waitForExplicitFinish();
        window.addEventListener("__taintreport", (report) => {

            is(report.detail.str, contents[i], "Check sink string content");

            let flow = report.detail.str.taint[0].flow;
            is(flow[2].operation, sink_names[i], `Check sink: ${sink_names[i]}`);

            i += 1;
            if (i >= sink_names.length) {
              SimpleTest.finish();
            }
        });

function secondBatch() {
              runTest2("json", r => { return r.response }, p => {
              SimpleTest.info(JSON.stringify(p));
              check_tainted(p.txt);
              let d = document.getElementById("foo");
              d.innerHTML = p.txt;
            }, "json");
            runTest2("xml", r => {return r.responseXML }, p => {
              SimpleTest.info(`XML: ${p.contentType}`);
              let txt = p.getElementById("bar").innerHTML;
              check_tainted(txt);
              let d = document.getElementById("foo");
              d.innerHTML = txt;
            }, "document");
            runTest2("html", r => r.response, p => {
              SimpleTest.info(`HTML: ${p.contentType}`);
              let txt = p.getElementById("bar").innerHTML;
              check_tainted(txt);
              let d = document.getElementById("foo");
              d.innerHTML = txt;
            }, "document");
}
            runTest("eval", r => r.responseText, p => eval(p));
            runTest("innerHTML", r => r.responseText, p => {
              let d = document.getElementById("foo");
              d.innerHTML = p;
            });
            let parser = new DOMParser();
            let d1 = parser.parseFromString("<html><body><div>", "text/html");

            runTest("doc_writeln", r => r.responseText, p => {
              d1.writeln(p);
            });

            runTest("doc_write", r => r.responseText, p => {
              d1.write(p);
              secondBatch();
            });




    </script>
  </head>

  <body>
    <div id="foo">

    </div>
</html>
